--- src/components/UserList/UserItem.module.css
@@ -1,72 +1,0 @@
-.UserItem {
-  background: white;
-  border-radius: 12px;
-  padding: 16px;
-  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.06);
-}
-.userItemContainer {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 16px; /* або скільки тобі треба */
-  padding: 12px 16px;
-  border-bottom: 1px solid #ccc;
-}
-.profileSection {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  flex: 1;
-}
-.profileImageContainer {
-  width: 48px;
-  height: 48px;
-  border-radius: 50%;
-  overflow: hidden;
-  background-color: #ccc;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.profileImage {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-.initialAvatar {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: white;
-  font-weight: bold;
-  font-size: 18px;
-  border-radius: 50%;
-}
-.profileName {
-  font-weight: 600;
-  color: #24448a;
-}
-.badgeAdmin {
-  margin-left: 6px;
-  font-size: 12px;
-  padding: 2px 6px;
-  background-color: #24448a;
-  color: white;
-  border-radius: 4px;
-}
-.iconButton {
-  background: none;
-  border: none;
-  cursor: pointer;
-  padding: 4px;
-}
-
-.emailSection,
-.timestampSection,
-.settingsSection,
-.actionsSection {
-  flex: 1;
-  text-align: center;
-}
