--- src/pages/UsersManagement/UsersManagement.module.css
@@ -1,5 +1,0 @@
-.usersPage {
-  padding: 52px 20px;
-  max-width: 1200px;
-  margin: 0 auto;
-}
