+++ src/pages/users/UsersManagement.module.css
@@ -1,0 +1,23 @@
+/* .usersPage {
+  padding: 52px 20px;
+  max-width: 1200px;
+  margin: 0 auto;
+} */
+.wrapper {
+  margin-top: 28px;
+}
+
+.backButton {
+  margin-bottom: 24px;
+}
+
+@media screen and (min-width: 768px) {
+  .wrapper {
+    margin-top: 49px;
+  }
+}
+@media screen and (min-width: 1440px) {
+  .wrapper {
+    margin-top: 56px;
+  }
+}
